Delete component integrated with modal
